+++ Tests/GRDBTests/DatabaseValueConvertibleFetchTests.swift
@@ -42,7 +42,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Fetched.fetchCursor(db, sql))
                 try test(Fetched.fetchCursor(statement))
                 try test(Fetched.fetchCursor(db, SQLRequest(sql)))
-                try test(SQLRequest(sql).fetching(Fetched.self).fetchCursor(db))
+                try test(SQLRequest(sql).asRequest(of: Fetched.self).fetchCursor(db))
             }
             do {
                 let sql = "SELECT 0, 1 UNION ALL SELECT 0, 2"
@@ -51,7 +51,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Fetched.fetchCursor(db, sql, adapter: adapter))
                 try test(Fetched.fetchCursor(statement, adapter: adapter))
                 try test(Fetched.fetchCursor(db, SQLRequest(sql, adapter: adapter)))
-                try test(SQLRequest(sql, adapter: adapter).fetching(Fetched.self).fetchCursor(db))
+                try test(SQLRequest(sql, adapter: adapter).asRequest(of: Fetched.self).fetchCursor(db))
             }
         }
     }
@@ -86,7 +86,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Fetched.fetchCursor(db, sql), sql: sql)
                 try test(Fetched.fetchCursor(db.makeSelectStatement(sql)), sql: sql)
                 try test(Fetched.fetchCursor(db, SQLRequest(sql)), sql: sql)
-                try test(SQLRequest(sql).fetching(Fetched.self).fetchCursor(db), sql: sql)
+                try test(SQLRequest(sql).asRequest(of: Fetched.self).fetchCursor(db), sql: sql)
             }
             do {
                 let sql = "SELECT 0, throw()"
@@ -94,7 +94,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Fetched.fetchCursor(db, sql, adapter: adapter), sql: sql)
                 try test(Fetched.fetchCursor(db.makeSelectStatement(sql), adapter: adapter), sql: sql)
                 try test(Fetched.fetchCursor(db, SQLRequest(sql, adapter: adapter)), sql: sql)
-                try test(SQLRequest(sql, adapter: adapter).fetching(Fetched.self).fetchCursor(db), sql: sql)
+                try test(SQLRequest(sql, adapter: adapter).asRequest(of: Fetched.self).fetchCursor(db), sql: sql)
             }
         }
     }
@@ -118,7 +118,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Fetched.fetchCursor(db, sql), sql: sql)
                 try test(Fetched.fetchCursor(db.makeSelectStatement(sql)), sql: sql)
                 try test(Fetched.fetchCursor(db, SQLRequest(sql)), sql: sql)
-                try test(SQLRequest(sql).fetching(Fetched.self).fetchCursor(db), sql: sql)
+                try test(SQLRequest(sql).asRequest(of: Fetched.self).fetchCursor(db), sql: sql)
             }
             do {
                 let sql = "SELECT * FROM nonExistingTable"
@@ -126,7 +126,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Fetched.fetchCursor(db, sql, adapter: adapter), sql: sql)
                 try test(Fetched.fetchCursor(db.makeSelectStatement(sql), adapter: adapter), sql: sql)
                 try test(Fetched.fetchCursor(db, SQLRequest(sql, adapter: adapter)), sql: sql)
-                try test(SQLRequest(sql, adapter: adapter).fetching(Fetched.self).fetchCursor(db), sql: sql)
+                try test(SQLRequest(sql, adapter: adapter).asRequest(of: Fetched.self).fetchCursor(db), sql: sql)
             }
         }
     }
@@ -143,7 +143,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Fetched.fetchAll(db, sql))
                 try test(Fetched.fetchAll(statement))
                 try test(Fetched.fetchAll(db, SQLRequest(sql)))
-                try test(SQLRequest(sql).fetching(Fetched.self).fetchAll(db))
+                try test(SQLRequest(sql).asRequest(of: Fetched.self).fetchAll(db))
             }
             do {
                 let sql = "SELECT 0, 1 UNION ALL SELECT 0, 2"
@@ -152,7 +152,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Fetched.fetchAll(db, sql, adapter: adapter))
                 try test(Fetched.fetchAll(statement, adapter: adapter))
                 try test(Fetched.fetchAll(db, SQLRequest(sql, adapter: adapter)))
-                try test(SQLRequest(sql, adapter: adapter).fetching(Fetched.self).fetchAll(db))
+                try test(SQLRequest(sql, adapter: adapter).asRequest(of: Fetched.self).fetchAll(db))
             }
         }
     }
@@ -178,7 +178,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Fetched.fetchAll(db, sql), sql: sql)
                 try test(Fetched.fetchAll(db.makeSelectStatement(sql)), sql: sql)
                 try test(Fetched.fetchAll(db, SQLRequest(sql)), sql: sql)
-                try test(SQLRequest(sql).fetching(Fetched.self).fetchAll(db), sql: sql)
+                try test(SQLRequest(sql).asRequest(of: Fetched.self).fetchAll(db), sql: sql)
             }
             do {
                 let sql = "SELECT 0, throw()"
@@ -186,7 +186,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Fetched.fetchAll(db, sql, adapter: adapter), sql: sql)
                 try test(Fetched.fetchAll(db.makeSelectStatement(sql), adapter: adapter), sql: sql)
                 try test(Fetched.fetchAll(db, SQLRequest(sql, adapter: adapter)), sql: sql)
-                try test(SQLRequest(sql, adapter: adapter).fetching(Fetched.self).fetchAll(db), sql: sql)
+                try test(SQLRequest(sql, adapter: adapter).asRequest(of: Fetched.self).fetchAll(db), sql: sql)
             }
         }
     }
@@ -210,7 +210,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Fetched.fetchAll(db, sql), sql: sql)
                 try test(Fetched.fetchAll(db.makeSelectStatement(sql)), sql: sql)
                 try test(Fetched.fetchAll(db, SQLRequest(sql)), sql: sql)
-                try test(SQLRequest(sql).fetching(Fetched.self).fetchAll(db), sql: sql)
+                try test(SQLRequest(sql).asRequest(of: Fetched.self).fetchAll(db), sql: sql)
             }
             do {
                 let sql = "SELECT * FROM nonExistingTable"
@@ -218,7 +218,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Fetched.fetchAll(db, sql, adapter: adapter), sql: sql)
                 try test(Fetched.fetchAll(db.makeSelectStatement(sql), adapter: adapter), sql: sql)
                 try test(Fetched.fetchAll(db, SQLRequest(sql, adapter: adapter)), sql: sql)
-                try test(SQLRequest(sql, adapter: adapter).fetching(Fetched.self).fetchAll(db), sql: sql)
+                try test(SQLRequest(sql, adapter: adapter).asRequest(of: Fetched.self).fetchAll(db), sql: sql)
             }
         }
     }
@@ -236,7 +236,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                     try test(Fetched.fetchOne(db, sql))
                     try test(Fetched.fetchOne(statement))
                     try test(Fetched.fetchOne(db, SQLRequest(sql)))
-                    try test(SQLRequest(sql).fetching(Fetched.self).fetchOne(db))
+                    try test(SQLRequest(sql).asRequest(of: Fetched.self).fetchOne(db))
                 }
                 do {
                     let sql = "SELECT 0, 1 WHERE 0"
@@ -245,7 +245,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                     try test(Fetched.fetchOne(db, sql, adapter: adapter))
                     try test(Fetched.fetchOne(statement, adapter: adapter))
                     try test(Fetched.fetchOne(db, SQLRequest(sql, adapter: adapter)))
-                    try test(SQLRequest(sql, adapter: adapter).fetching(Fetched.self).fetchOne(db))
+                    try test(SQLRequest(sql, adapter: adapter).asRequest(of: Fetched.self).fetchOne(db))
                 }
             }
             do {
@@ -258,7 +258,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                     try test(Fetched.fetchOne(db, sql))
                     try test(Fetched.fetchOne(statement))
                     try test(Fetched.fetchOne(db, SQLRequest(sql)))
-                    try test(SQLRequest(sql).fetching(Fetched.self).fetchOne(db))
+                    try test(SQLRequest(sql).asRequest(of: Fetched.self).fetchOne(db))
                 }
                 do {
                     let sql = "SELECT 0, NULL"
@@ -267,7 +267,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                     try test(Fetched.fetchOne(db, sql, adapter: adapter))
                     try test(Fetched.fetchOne(statement, adapter: adapter))
                     try test(Fetched.fetchOne(db, SQLRequest(sql, adapter: adapter)))
-                    try test(SQLRequest(sql, adapter: adapter).fetching(Fetched.self).fetchOne(db))
+                    try test(SQLRequest(sql, adapter: adapter).asRequest(of: Fetched.self).fetchOne(db))
                 }
             }
             do {
@@ -280,7 +280,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                     try test(Fetched.fetchOne(db, sql))
                     try test(Fetched.fetchOne(statement))
                     try test(Fetched.fetchOne(db, SQLRequest(sql)))
-                    try test(SQLRequest(sql).fetching(Fetched.self).fetchOne(db))
+                    try test(SQLRequest(sql).asRequest(of: Fetched.self).fetchOne(db))
                 }
                 do {
                     let sql = "SELECT 0, 1"
@@ -289,7 +289,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                     try test(Fetched.fetchOne(db, sql, adapter: adapter))
                     try test(Fetched.fetchOne(statement, adapter: adapter))
                     try test(Fetched.fetchOne(db, SQLRequest(sql, adapter: adapter)))
-                    try test(SQLRequest(sql, adapter: adapter).fetching(Fetched.self).fetchOne(db))
+                    try test(SQLRequest(sql, adapter: adapter).asRequest(of: Fetched.self).fetchOne(db))
                 }
             }
         }
@@ -316,7 +316,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Fetched.fetchOne(db, sql), sql: sql)
                 try test(Fetched.fetchOne(db.makeSelectStatement(sql)), sql: sql)
                 try test(Fetched.fetchOne(db, SQLRequest(sql)), sql: sql)
-                try test(SQLRequest(sql).fetching(Fetched.self).fetchOne(db), sql: sql)
+                try test(SQLRequest(sql).asRequest(of: Fetched.self).fetchOne(db), sql: sql)
             }
             do {
                 let sql = "SELECT 0, throw()"
@@ -324,7 +324,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Fetched.fetchOne(db, sql, adapter: adapter), sql: sql)
                 try test(Fetched.fetchOne(db.makeSelectStatement(sql), adapter: adapter), sql: sql)
                 try test(Fetched.fetchOne(db, SQLRequest(sql, adapter: adapter)), sql: sql)
-                try test(SQLRequest(sql, adapter: adapter).fetching(Fetched.self).fetchOne(db), sql: sql)
+                try test(SQLRequest(sql, adapter: adapter).asRequest(of: Fetched.self).fetchOne(db), sql: sql)
             }
         }
     }
@@ -348,7 +348,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Fetched.fetchOne(db, sql), sql: sql)
                 try test(Fetched.fetchOne(db.makeSelectStatement(sql)), sql: sql)
                 try test(Fetched.fetchOne(db, SQLRequest(sql)), sql: sql)
-                try test(SQLRequest(sql).fetching(Fetched.self).fetchOne(db), sql: sql)
+                try test(SQLRequest(sql).asRequest(of: Fetched.self).fetchOne(db), sql: sql)
             }
             do {
                 let sql = "SELECT * FROM nonExistingTable"
@@ -356,7 +356,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Fetched.fetchOne(db, sql, adapter: adapter), sql: sql)
                 try test(Fetched.fetchOne(db.makeSelectStatement(sql), adapter: adapter), sql: sql)
                 try test(Fetched.fetchOne(db, SQLRequest(sql, adapter: adapter)), sql: sql)
-                try test(SQLRequest(sql, adapter: adapter).fetching(Fetched.self).fetchOne(db), sql: sql)
+                try test(SQLRequest(sql, adapter: adapter).asRequest(of: Fetched.self).fetchOne(db), sql: sql)
             }
         }
     }
@@ -377,7 +377,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Optional<Fetched>.fetchCursor(db, sql))
                 try test(Optional<Fetched>.fetchCursor(statement))
                 try test(Optional<Fetched>.fetchCursor(db, SQLRequest(sql)))
-                try test(SQLRequest(sql).fetching(Optional<Fetched>.self).fetchCursor(db))
+                try test(SQLRequest(sql).asRequest(of: Optional<Fetched>.self).fetchCursor(db))
             }
             do {
                 let sql = "SELECT 0, 1 UNION ALL SELECT 0, NULL"
@@ -386,7 +386,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Optional<Fetched>.fetchCursor(db, sql, adapter: adapter))
                 try test(Optional<Fetched>.fetchCursor(statement, adapter: adapter))
                 try test(Optional<Fetched>.fetchCursor(db, SQLRequest(sql, adapter: adapter)))
-                try test(SQLRequest(sql, adapter: adapter).fetching(Optional<Fetched>.self).fetchCursor(db))
+                try test(SQLRequest(sql, adapter: adapter).asRequest(of: Optional<Fetched>.self).fetchCursor(db))
             }
         }
     }
@@ -421,7 +421,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Optional<Fetched>.fetchCursor(db, sql), sql: sql)
                 try test(Optional<Fetched>.fetchCursor(db.makeSelectStatement(sql)), sql: sql)
                 try test(Optional<Fetched>.fetchCursor(db, SQLRequest(sql)), sql: sql)
-                try test(SQLRequest(sql).fetching(Optional<Fetched>.self).fetchCursor(db), sql: sql)
+                try test(SQLRequest(sql).asRequest(of: Optional<Fetched>.self).fetchCursor(db), sql: sql)
             }
             do {
                 let sql = "SELECT 0, throw()"
@@ -429,7 +429,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Optional<Fetched>.fetchCursor(db, sql, adapter: adapter), sql: sql)
                 try test(Optional<Fetched>.fetchCursor(db.makeSelectStatement(sql), adapter: adapter), sql: sql)
                 try test(Optional<Fetched>.fetchCursor(db, SQLRequest(sql, adapter: adapter)), sql: sql)
-                try test(SQLRequest(sql, adapter: adapter).fetching(Optional<Fetched>.self).fetchCursor(db), sql: sql)
+                try test(SQLRequest(sql, adapter: adapter).asRequest(of: Optional<Fetched>.self).fetchCursor(db), sql: sql)
             }
         }
     }
@@ -453,7 +453,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Optional<Fetched>.fetchCursor(db, sql), sql: sql)
                 try test(Optional<Fetched>.fetchCursor(db.makeSelectStatement(sql)), sql: sql)
                 try test(Optional<Fetched>.fetchCursor(db, SQLRequest(sql)), sql: sql)
-                try test(SQLRequest(sql).fetching(Optional<Fetched>.self).fetchCursor(db), sql: sql)
+                try test(SQLRequest(sql).asRequest(of: Optional<Fetched>.self).fetchCursor(db), sql: sql)
             }
             do {
                 let sql = "SELECT * FROM nonExistingTable"
@@ -461,7 +461,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Optional<Fetched>.fetchCursor(db, sql, adapter: adapter), sql: sql)
                 try test(Optional<Fetched>.fetchCursor(db.makeSelectStatement(sql), adapter: adapter), sql: sql)
                 try test(Optional<Fetched>.fetchCursor(db, SQLRequest(sql, adapter: adapter)), sql: sql)
-                try test(SQLRequest(sql, adapter: adapter).fetching(Optional<Fetched>.self).fetchCursor(db), sql: sql)
+                try test(SQLRequest(sql, adapter: adapter).asRequest(of: Optional<Fetched>.self).fetchCursor(db), sql: sql)
             }
         }
     }
@@ -480,7 +480,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Optional<Fetched>.fetchAll(db, sql))
                 try test(Optional<Fetched>.fetchAll(statement))
                 try test(Optional<Fetched>.fetchAll(db, SQLRequest(sql)))
-                try test(SQLRequest(sql).fetching(Optional<Fetched>.self).fetchAll(db))
+                try test(SQLRequest(sql).asRequest(of: Optional<Fetched>.self).fetchAll(db))
             }
             do {
                 let sql = "SELECT 0, 1 UNION ALL SELECT 0, NULL"
@@ -489,7 +489,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Optional<Fetched>.fetchAll(db, sql, adapter: adapter))
                 try test(Optional<Fetched>.fetchAll(statement, adapter: adapter))
                 try test(Optional<Fetched>.fetchAll(db, SQLRequest(sql, adapter: adapter)))
-                try test(SQLRequest(sql, adapter: adapter).fetching(Optional<Fetched>.self).fetchAll(db))
+                try test(SQLRequest(sql, adapter: adapter).asRequest(of: Optional<Fetched>.self).fetchAll(db))
             }
         }
     }
@@ -515,7 +515,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Optional<Fetched>.fetchAll(db, sql), sql: sql)
                 try test(Optional<Fetched>.fetchAll(db.makeSelectStatement(sql)), sql: sql)
                 try test(Optional<Fetched>.fetchAll(db, SQLRequest(sql)), sql: sql)
-                try test(SQLRequest(sql).fetching(Optional<Fetched>.self).fetchAll(db), sql: sql)
+                try test(SQLRequest(sql).asRequest(of: Optional<Fetched>.self).fetchAll(db), sql: sql)
             }
             do {
                 let sql = "SELECT 0, throw()"
@@ -523,7 +523,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Optional<Fetched>.fetchAll(db, sql, adapter: adapter), sql: sql)
                 try test(Optional<Fetched>.fetchAll(db.makeSelectStatement(sql), adapter: adapter), sql: sql)
                 try test(Optional<Fetched>.fetchAll(db, SQLRequest(sql, adapter: adapter)), sql: sql)
-                try test(SQLRequest(sql, adapter: adapter).fetching(Optional<Fetched>.self).fetchAll(db), sql: sql)
+                try test(SQLRequest(sql, adapter: adapter).asRequest(of: Optional<Fetched>.self).fetchAll(db), sql: sql)
             }
         }
     }
@@ -547,7 +547,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Optional<Fetched>.fetchAll(db, sql), sql: sql)
                 try test(Optional<Fetched>.fetchAll(db.makeSelectStatement(sql)), sql: sql)
                 try test(Optional<Fetched>.fetchAll(db, SQLRequest(sql)), sql: sql)
-                try test(SQLRequest(sql).fetching(Optional<Fetched>.self).fetchAll(db), sql: sql)
+                try test(SQLRequest(sql).asRequest(of: Optional<Fetched>.self).fetchAll(db), sql: sql)
             }
             do {
                 let sql = "SELECT * FROM nonExistingTable"
@@ -555,7 +555,7 @@ class DatabaseValueConvertibleFetchTests: GRDBTestCase {
                 try test(Optional<Fetched>.fetchAll(db, sql, adapter: adapter), sql: sql)
                 try test(Optional<Fetched>.fetchAll(db.makeSelectStatement(sql), adapter: adapter), sql: sql)
                 try test(Optional<Fetched>.fetchAll(db, SQLRequest(sql, adapter: adapter)), sql: sql)
-                try test(SQLRequest(sql, adapter: adapter).fetching(Optional<Fetched>.self).fetchAll(db), sql: sql)
+                try test(SQLRequest(sql, adapter: adapter).asRequest(of: Optional<Fetched>.self).fetchAll(db), sql: sql)
             }
         }
     }