+++ Tests/GRDBTests/RowAdapterTests.swift
@@ -668,7 +668,7 @@ class AdapterRowTests : RowTestCase {
         try dbQueue.inDatabase { db in
             do {
                 let row = try SQLRequest("SELECT 0 AS a0, 1 AS a1, 2 AS a2", adapter: SuffixRowAdapter(fromIndex: 1))
-                    .fetching(Row.self)
+                    .asRequest(of: Row.self)
                     .fetchOne(db)!
                 XCTAssertEqual(Array(row.columnNames), ["a1", "a2"])
                 XCTAssertEqual(Array(row.databaseValues), [1.databaseValue, 2.databaseValue])
@@ -676,7 +676,7 @@ class AdapterRowTests : RowTestCase {
             do {
                 let row = try SQLRequest("SELECT 0 AS a0, 1 AS a1, 2 AS a2")
                     .adapted { _ in SuffixRowAdapter(fromIndex: 1) }
-                    .fetching(Row.self)
+                    .asRequest(of: Row.self)
                     .fetchOne(db)!
                 XCTAssertEqual(Array(row.columnNames), ["a1", "a2"])
                 XCTAssertEqual(Array(row.databaseValues), [1.databaseValue, 2.databaseValue])
@@ -684,7 +684,7 @@ class AdapterRowTests : RowTestCase {
             do {
                 let row = try SQLRequest("SELECT 0 AS a0, 1 AS a1, 2 AS a2", adapter: SuffixRowAdapter(fromIndex: 1))
                     .adapted { _ in SuffixRowAdapter(fromIndex: 1) }
-                    .fetching(Row.self)
+                    .asRequest(of: Row.self)
                     .fetchOne(db)!
                 XCTAssertEqual(Array(row.columnNames), ["a2"])
                 XCTAssertEqual(Array(row.databaseValues), [2.databaseValue])
@@ -692,7 +692,7 @@ class AdapterRowTests : RowTestCase {
             do {
                 let row = try SQLRequest("SELECT 0 AS a0", adapter: ColumnMapping(["a1": "a0"]))
                     .adapted { _ in ColumnMapping(["a2": "a1"]) }
-                    .fetching(Row.self)
+                    .asRequest(of: Row.self)
                     .fetchOne(db)!
                 XCTAssertEqual(Array(row.columnNames), ["a2"])
                 XCTAssertEqual(Array(row.databaseValues), [0.databaseValue])
@@ -705,14 +705,14 @@ class AdapterRowTests : RowTestCase {
         try dbQueue.inDatabase { db in
             do {
                 let row = try SQLRequest("SELECT 0 AS a0, 1 AS a1, 2 AS a2", adapter: SuffixRowAdapter(fromIndex: 1))
-                    .fetching(Row.self)
+                    .asRequest(of: Row.self)
                     .fetchOne(db)!
                 XCTAssertEqual(Array(row.columnNames), ["a1", "a2"])
                 XCTAssertEqual(Array(row.databaseValues), [1.databaseValue, 2.databaseValue])
             }
             do {
                 let row = try SQLRequest("SELECT 0 AS a0, 1 AS a1, 2 AS a2")
-                    .fetching(Row.self)
+                    .asRequest(of: Row.self)
                     .adapted { _ in SuffixRowAdapter(fromIndex: 1) }
                     .fetchOne(db)!
                 XCTAssertEqual(Array(row.columnNames), ["a1", "a2"])
@@ -720,7 +720,7 @@ class AdapterRowTests : RowTestCase {
             }
             do {
                 let row = try SQLRequest("SELECT 0 AS a0, 1 AS a1, 2 AS a2", adapter: SuffixRowAdapter(fromIndex: 1))
-                    .fetching(Row.self)
+                    .asRequest(of: Row.self)
                     .adapted { _ in SuffixRowAdapter(fromIndex: 1) }
                     .fetchOne(db)!
                 XCTAssertEqual(Array(row.columnNames), ["a2"])
@@ -728,7 +728,7 @@ class AdapterRowTests : RowTestCase {
             }
             do {
                 let row = try SQLRequest("SELECT 0 AS a0", adapter: ColumnMapping(["a1": "a0"]))
-                    .fetching(Row.self)
+                    .asRequest(of: Row.self)
                     .adapted { _ in ColumnMapping(["a2": "a1"]) }
                     .fetchOne(db)!
                 XCTAssertEqual(Array(row.columnNames), ["a2"])